changes condition status to be a string, rather than a boolean
Closes #13074
