+++ operator/src/main/java/org/keycloak/operator/crds/v2alpha1/realmimport/KeycloakRealmImportStatusCondition.java
@@ -17,61 +17,10 @@
 
 package org.keycloak.operator.crds.v2alpha1.realmimport;
 
-import java.util.Objects;
+import org.keycloak.operator.crds.v2alpha1.StatusCondition;
 
-public class KeycloakRealmImportStatusCondition {
+public class KeycloakRealmImportStatusCondition extends StatusCondition {
     public static final String DONE = "Done";
     public static final String STARTED = "Started";
     public static final String HAS_ERRORS = "HasErrors";
-
-    // string to avoid enums in CRDs
-    private String type;
-    private Boolean status;
-    private String message;
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public Boolean getStatus() {
-        return status;
-    }
-
-    public void setStatus(Boolean status) {
-        this.status = status;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        KeycloakRealmImportStatusCondition that = (KeycloakRealmImportStatusCondition) o;
-        return getType() == that.getType() && Objects.equals(getStatus(), that.getStatus()) && Objects.equals(getMessage(), that.getMessage());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getType(), getStatus(), getMessage());
-    }
-
-    @Override
-    public String toString() {
-        return "KeycloakRealmImportStatusCondition{" +
-                "type='" + type + '\'' +
-                ", status=" + status +
-                ", message='" + message + '\'' +
-                '}';
-    }
 }
